Move lexer tests to a separate file
I just feel somewhat uneasy about 3 KLOCs of unit-tests right after
the source code. These are more of a conformance tests rather than
example-of-use unit-test-like tests, so they should be kept aside.
So I simply moved them out into 'tests' subdirectory, Cargo will
compile and run them anyway.
